ingestion/image_description/requirements.txt
google-cloud-storage==2.17.0
google-cloud-bigquery==3.25.0
google-cloud-logging==3.11.1
vertexai==1.49.0
PyMuPDF==1.24.7
gcsfs==2024.6.1
pytz==2024.1
fastapi==0.112.1
uvicorn==0.30.6